BPE26-ref: extract fullscreen feature to dedicated file
